+++ service/controller/resource/monitoring/prometheus/resource.go
@@ -124,7 +124,7 @@ func toPrometheus(ctx context.Context, v interface{}, config Config) (metav1.Obj
 		VolumeClaimTemplate: promv1.EmbeddedPersistentVolumeClaim{
 			Spec: corev1.PersistentVolumeClaimSpec{
 				AccessModes: []corev1.PersistentVolumeAccessMode{corev1.ReadWriteOnce},
-				Resources: corev1.ResourceRequirements{
+				Resources: corev1.VolumeResourceRequirements{
 					Requests: corev1.ResourceList{
 						corev1.ResourceStorage: storageSize,
 					},